split sensibly
